+++ src/authorization.ts
@@ -1,9 +1,12 @@
+import { CreativeWorkType } from './creativeWorkType';
+import { IMember, IMemberOfRole } from './iam';
 import { IOrder } from './order';
+import { OrganizationType } from './organizationType';
 import { IOwnershipInfo, IPermitAsGood, IPermitIssuedThrough, IReservation } from './ownershipInfo';
+import { PersonType } from './personType';
 import { IProject } from './project';
 import { SortType } from './sortType';
 
-// strict definetion(2023-01-31~)
 export type IOrderAsObject = Pick<IOrder, 'orderNumber' | 'typeOf'>;
 export type IPermitIssuedThroughOfObjectTypeOfGood = Pick<IPermitIssuedThrough, 'id' | 'typeOf'>;
 export type IPermitAsObjectTypeOfGood = Pick<IPermitAsGood, 'identifier' | 'typeOf'> & {
@@ -12,7 +15,28 @@ export type IPermitAsObjectTypeOfGood = Pick<IPermitAsGood, 'identifier' | 'type
 export type IPermitOwnershipInfoAsObject = Pick<IOwnershipInfo<IPermitAsObjectTypeOfGood>, 'id' | 'identifier' | 'typeOf' | 'typeOfGood'>;
 export type IReservationOwnershipInfoAsObject = Pick<IOwnershipInfo<IReservation>, 'id' | 'identifier' | 'typeOf' | 'typeOfGood'>;
 export type IOwnershipInfoAsObject = IPermitOwnershipInfoAsObject | IReservationOwnershipInfoAsObject;
-export type IObject = IOrderAsObject | IOwnershipInfoAsObject;
+/**
+ * 承認対象としてのIAMメンバー
+ */
+export type IRoleAsObject = Pick<IMember, 'member' | 'typeOf'> & {
+    member: Pick<IMemberOfRole, 'hasRole' | 'id' | 'memberOf' | 'typeOf'>;
+};
+export type IObject = IOrderAsObject | IOwnershipInfoAsObject | IRoleAsObject;
+export interface IAudience {
+    /**
+     * クライアントID
+     */
+    id: string;
+    typeOf: CreativeWorkType.SoftwareApplication | CreativeWorkType.WebApplication;
+}
+export interface IAuthor {
+    id: string;
+    typeOf: PersonType.Person | CreativeWorkType.WebApplication;
+}
+export interface IIssuedBy {
+    id: string;
+    typeOf: OrganizationType.Corporation | OrganizationType.Project;
+}
 
 /**
  * 承認
@@ -20,6 +44,7 @@ export type IObject = IOrderAsObject | IOwnershipInfoAsObject;
 export interface IAuthorization {
     project: Pick<IProject, 'id' | 'typeOf'>;
     typeOf: 'Authorization';
+    id?: string;
     /**
      * 承認コード
      */
@@ -36,6 +61,9 @@ export interface IAuthorization {
      * 有効期間
      */
     validUntil: Date;
+    audience?: IAudience;
+    author?: IAuthor;
+    issuedBy?: IIssuedBy;
 }
 
 /**