+++ src/index.ts
@@ -60,6 +60,7 @@ import * as CommentFactory from './creativeWork/comment';
 import * as EmailMessageFactory from './creativeWork/message/email';
 import * as MovieFactory from './creativeWork/movie';
 import * as NoteDigitalDocumentFactory from './creativeWork/noteDigitalDocument';
+import * as SoftwareApplicationFactory from './creativeWork/softwareApplication';
 import * as WebApplicationFactory from './creativeWork/softwareApplication/webApplication';
 import { CreativeWorkType } from './creativeWorkType';
 import * as CustomerFactory from './customer';
@@ -351,6 +352,7 @@ export namespace creativeWork {
     export import movie = MovieFactory;
     export import noteDigitalDocument = NoteDigitalDocumentFactory;
     export namespace softwareApplication {
+        export import ISoftwareApplication = SoftwareApplicationFactory.ISoftwareApplication;
         export import webApplication = WebApplicationFactory;
     }
 }